% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{db_connect}
\alias{db_connect}
\title{Connect to a HUD database}
\usage{
db_connect(
  database = "property",
  server = "property.database.windows.net",
  driver = "{ODBC Driver 18 for SQL Server}"
)
}
\arguments{
\item{database}{Database name (optional)}

\item{server}{Server name (optional)}

\item{driver}{Driver string (optional)}
}
\description{
Handles all the Active Directory authentication and connects to database.
}
