% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{find_latest}
\alias{find_latest}
\title{Find latest}
\usage{
find_latest(blob_pattern, container_url, dir)
}
\arguments{
\item{blob_pattern}{Pattern to match for (e.g. "ea-icp-grab.csv")}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{dir}{Base path to search in (e.g. "hmu-bot")}
}
\description{
Finds the latest version of a file stored on the blob. The last file,
sorted by full blob name, is considered the latest.
}
\details{
e.g.:

"hmu-bot/hmu-bot-20230115/Construction/ea-icp-grab.csv"
"hmu-bot/hmu-bot-20230117/Construction/ea-icp-grab.csv"
"hmu-bot/hmu-bot-20230120/Construction/ea-icp-grab.csv" <- Latest
}
