% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{get_container}
\alias{get_container}
\title{Get container}
\usage{
get_container(container_url)
}
\arguments{
\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}
}
\description{
Performs authentication and obtains blob container object.
}
