% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_query}
\alias{get_query}
\alias{get_spatial_query}
\title{Run query}
\usage{
get_query(query, ...)

get_spatial_query(query, geometry_column = "geom", ...)
}
\arguments{
\item{query}{SQL query}

\item{geometry_column}{Specific the column for sf to treat as geometry}

\item{database}{Database name (optional)}

\item{server}{Server name (optional)}

\item{driver}{Driver string (optional)}
}
\description{
Shortcut for running a SQL query

Shortcut for running a SQL query and cleaning spatial data
}
