% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{list_local}
\alias{list_local}
\title{List local}
\usage{
list_local(trg_path, full.names = FALSE)
}
\arguments{
\item{trg_path}{Path}

\item{full.names}{Whether to include source path in the file names}
}
\description{
List all the files stored within a local folder, with hashes.
}
