% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{list_stored}
\alias{list_stored}
\title{List stored}
\usage{
list_stored(blob_starts_with, container_url)
}
\arguments{
\item{blob_starts_with}{Path or prefix to look for}

\item{container_url}{Azure container URL (e.g. "https://sysintel.blob.core.windows.net/bot-outputs")}
}
\description{
List all the files stored in the blob.
}
