% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{read_blob}
\alias{read_blob}
\title{DEPRECATED - Read blob data file}
\usage{
read_blob(blob_fn, container_url, sheet = NULL, ...)
}
\arguments{
\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{sheet}{Sheet name for Excel}
}
\description{
DEPRECATED - Read a CSV/Excel file from the blob
}
