% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{read_blob_using}
\alias{read_blob_using}
\title{Read blob file using custom function}
\usage{
read_blob_using(blob_fn, container_url, f, forced = FALSE, ...)
}
\arguments{
\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{f}{Custom function}

\item{forced}{Overwrite local version}
}
\description{
Read a file from the blob using custom function
}
