% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{store}
\alias{store}
\title{Store}
\usage{
store(local_fn, blob_fn, container_url, forced = FALSE)
}
\arguments{
\item{local_fn}{Local filename}

\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlreportingdataprod.blob.core.windows.net/bot-outputs")}

\item{forced}{Overwrite blob version}
}
\description{
Stores a local file in the blob.
}
