% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{store_folder}
\alias{store_folder}
\title{Store folder}
\usage{
store_folder(
  local_path,
  blob_path,
  container_url,
  update = TRUE,
  forced = FALSE
)
}
\arguments{
\item{local_path}{Local folder (e.g. "outputs/20250101")}

\item{blob_path}{Blob path where the contents of the local folder will be uploaded to (e.g. "project_name/outputs/20250101")}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{update}{Update the blob version if the local version is newer}

\item{forced}{Overwrite blob version no matter what}
}
\description{
Recursively store everything inside a folder in the blob.
}
