% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{read_blob_data}
\alias{read_blob_data}
\title{Read blob data file}
\usage{
read_blob_data(blob_fn, container_url, forced = FALSE, ...)
}
\arguments{
\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{forced}{Overwrite local version}
}
\description{
Read a CSV/Excel file from the blob
}
