% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{read_blob_using}
\alias{read_blob_using}
\title{\link{DEPRECATED} Read blob file using custom function}
\usage{
read_blob_using(blob_fn, container_url, f = NA, ...)
}
\arguments{
\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{f}{Custom function}

\item{...}{Additional parameters for custom function}
}
\description{
Use read_blob_data() instead
}
