% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper.R
\name{retrieve}
\alias{retrieve}
\title{Retrive}
\usage{
retrieve(blob_fn, local_fn, container_url, forced = FALSE)
}
\arguments{
\item{blob_fn}{Blob filename (including path)}

\item{local_fn}{Local filename}

\item{container_url}{Azure container URL (e.g. "https://dlreportingdataprod.blob.core.windows.net/bot-outputs")}

\item{forced}{Overwrite local version}
}
\description{
Retrives a file from the blob and save it locally.
}
