% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{store}
\alias{store}
\title{Store}
\usage{
store(local_fn, blob_fn, container_url, update = TRUE, forced = FALSE)
}
\arguments{
\item{local_fn}{Local filename}

\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{update}{Update the blob version if the local version is newer}

\item{forced}{Overwrite blob version no matter what}
}
\description{
Stores a local file in the blob.
}
