% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{store_data}
\alias{store_data}
\title{Store data}
\usage{
store_data(x, blob_fn, container_url, f = saveRDS, forced = FALSE)
}
\arguments{
\item{x}{Dataframe or tibble}

\item{blob_fn}{Blob filename (including path)}

\item{container_url}{Azure container URL (e.g. "https://dlprojectsdataprod.blob.core.windows.net/bot-outputs")}

\item{forced}{Overwrite blob version}

\item{file_format}{Format for saving the file}
}
\description{
Stores a dataframe or tibble in the blob.
}
